Add a separate client class file
