+++ gapic-generator-cloud/templates/cloud/client/_new.erb
@@ -1,3 +1,7 @@
+<%- assert_locals service -%>
+<% @requires = capture do %>
+require "<%= service.client_class_require %>"
+<% end %>
 ##
 # @param credentials [Google::Auth::Credentials, String, Hash,
 #   GRPC::Core::Channel, GRPC::Core::ChannelCredentials, Proc]
