+++ gapic-generator-cloud/templates/cloud/client/class.erb
@@ -1,5 +1,5 @@
 <%- assert_locals service -%>
-<%= render partial: "client/new",
+<%= render partial: "client/class",
            layout: "layouts/ruby",
            locals: { service: service,
                      namespaces: service.namespaces }
