+++ gapic-generator-cloud/templates/cloud/client/credentials.erb
